+++ doorstop/core/publisher.py
@@ -7,7 +7,7 @@ import logging
 import markdown
 
 from doorstop.common import DoorstopError, create_dirname
-from doorstop.core.types import iter_documents, iter_items, is_tree
+from doorstop.core.types import iter_documents, iter_items, is_tree, is_item
 from doorstop import settings
 
 CSS = os.path.join(os.path.dirname(__file__), 'files', 'doorstop.css')
@@ -30,6 +30,8 @@ def publish(obj, path, ext=None, linkify=None, index=None, **kwargs):
 
     @raise DoorstopError: for unknown file formats
 
+    @return: output location if files created, else None
+
     """
     # Determine the output format
     ext = ext or os.path.splitext(path)[-1] or '.html'
@@ -38,18 +40,30 @@ def publish(obj, path, ext=None, linkify=None, index=None, **kwargs):
     check(ext)
 
     # Publish documents
+    count = 0
     for obj2, path2 in iter_documents(obj, path, ext):
+        count += 1
 
         # Publish content to the specified path
         create_dirname(path2)
         logging.info("creating file {}...".format(path2))
-        with open(path2, 'w') as outfile:  # pragma: no cover (integration test)
+        with open(path2, 'wb') as outfile:  # pragma: no cover (integration test)
             for line in publish_lines(obj2, ext, linkify=linkify, **kwargs):
-                outfile.write(line + '\n')
+                outfile.write((line + '\n').encode('utf-8'))
 
     # Create index
-    if index:
+    if index and count:
+        count += 1
         _index(path)
+
+    # Return the published path
+    if count:
+        msg = "created {} file{}".format(count, 's' if count > 1 else '')
+        logging.info(msg)
+        return path
+    else:
+        logging.warning("nothing to publish")
+        return None
 
 
 def _index(directory, extensions=('.html',)):
@@ -262,14 +276,15 @@ def _format_ref(item):
 
 def _format_links(items, linkify):
     """Format a list of linked items."""
-    if linkify:
-        links = []
-        for item in items:
-            links.append("[{i}]({p}.html#{i})".format(i=item.id,
-                                                      p=item.document.prefix))
-        return ', '.join(links)
-    else:
-        return ', '.join(str(item.id) for item in items)
+    links = []
+    for item in items:
+        if is_item(item) and linkify:
+            link = "[{i}]({p}.html#{i})".format(i=item.id,
+                                                p=item.document.prefix)
+        else:
+            link = str(item.id)  # assume this is an `UnknownItem`
+        links.append(link)
+    return ', '.join(links)
 
 
 def _format_label_links(label, links, linkify):