+++ doorstop/core/types.py
@@ -15,6 +15,8 @@ class Prefix(str):  # pylint: disable=R0904
 
     """Unique document prefixes."""
 
+    UNKNOWN_MESSGE = "no document with prefix: {}"
+
     def __new__(cls, value=""):
         if isinstance(value, Prefix):
             return value
@@ -54,6 +56,8 @@ class Prefix(str):  # pylint: disable=R0904
 class ID(object):
 
     """Unique item identifier."""
+
+    UNKNOWN_MESSAGE = "no{k} item with ID: {i}"  # k='parent'|'child'|'', i=ID
 
     def __new__(cls, *values):
         if values and isinstance(values[0], ID):
@@ -347,7 +351,8 @@ class Level(object):
                 self._parts = parts
                 self.heading = False
         self.heading = self.heading if heading is None else heading
-        self._adjust()
+        if not value:
+            self._adjust()
 
     def __repr__(self):
         if self.heading:
@@ -538,14 +543,33 @@ class Level(object):
         return Level(self.value)
 
 
+def is_tree(obj):
+    """Determine if the object is a tree-like."""
+    return hasattr(obj, 'documents')
+
+
 def is_document(obj):
-    """Determine if the object is a document."""
+    """Determine if the object is a document-like."""
     return hasattr(obj, 'items')
 
 
-def is_tree(obj):
-    """Determine if the object is a tree."""
-    return hasattr(obj, 'documents')
+def is_item(obj):
+    """Determine if the object is item-like."""
+    return hasattr(obj, 'text')
+
+
+def iter_documents(obj, path, ext):
+    """Get an iterator if documents from a tree or document-like object."""
+    if is_tree(obj):
+        # a tree
+        logging.debug("iterating over tree...")
+        for document in obj.documents:
+            path2 = os.path.join(path, document.prefix + ext)
+            yield document, path2
+    else:
+        # assume a document-like object
+        logging.debug("iterating over document-like object...")
+        yield obj, path
 
 
 def iter_items(obj):
@@ -562,17 +586,3 @@ def iter_items(obj):
         # an item
         logging.debug("iterating over an item (in a container)...")
         return [obj]
-
-
-def iter_documents(obj, path, ext):
-    """Get an iterator if documents from a tree or document-like object."""
-    if is_tree(obj):
-        # a tree
-        logging.debug("iterating over tree...")
-        for document in obj.documents:
-            path2 = os.path.join(path, document.prefix + ext)
-            yield document, path2
-    else:
-        # assume a document-like object
-        logging.debug("iterating over document-like object...")
-        yield obj, path